separated modules and organized them
