+++ src/assets/explosion_somewhere_far.mp3
