+++ src/assets/sprBg1.png
