+++ src/assets/sprBtnPlay.png
