+++ src/assets/sprBtnRestart.png
