+++ src/entities/ChaserShip.js
@@ -1,0 +1,38 @@
+import {Entity} from './Entities'
+export class ChaserShip extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, "ufo", "ChaserShip");
+    this.body.velocity.y = Phaser.Math.Between(50, 100);
+    this.states = {
+      MOVE_DOWN: "MOVE_DOWN",
+      CHASE: "CHASE"
+    };
+    this.state = this.states.MOVE_DOWN;
+  }
+  update(){
+    if (!this.getData("isDead") && this.scene.player) {
+      if (Phaser.Math.Distance.Between(
+        this.x,
+        this.y,
+        this.scene.player.x,
+        this.scene.player.y
+      ) < 320) {
+
+        this.state = this.states.CHASE;
+      }
+
+      if (this.state == this.states.CHASE) {
+        var dx = this.scene.player.x - this.x;
+        var dy = this.scene.player.y - this.y;
+
+        var angle = Math.atan2(dy, dx);
+
+        var speed = 100;
+        this.body.setVelocity(
+          Math.cos(angle) * speed,
+          Math.sin(angle) * speed
+        );
+      }
+    }
+  }
+}
