+++ src/entities/Entities.js
@@ -1,0 +1,34 @@
+export class Entity extends Phaser.GameObjects.Sprite {
+  constructor(scene, x, y, key, type) {
+    super(scene, x, y, key);
+
+    this.scene = scene;
+    this.scene.add.existing(this);
+    this.scene.physics.world.enableBody(this, 0);
+    this.setData("type", type);
+    this.setData("isDead", false);
+  }
+  explode(canDestroy) {
+    if (!this.getData("isDead")) {
+      this.setTexture("exp2_0");  // this refers to the same animation key we used when we added this.anims.create previously
+      this.play("exp2_0");
+      this.scene.sfx.explosions.play();
+      if (this.shootTimer !== undefined) {
+        if (this.shootTimer) {
+          this.shootTimer.remove(false);
+        }
+      }
+      this.setAngle(0);
+      this.body.setVelocity(0, 0);
+      this.on('animationcomplete', function() {
+        if (canDestroy) {
+          this.destroy();
+        }
+        else {
+          this.setVisible(false);
+        }
+      }, this);
+      this.setData("isDead", true);
+    }
+  }
+}
