+++ src/entities/PlayerLaser.js
@@ -1,0 +1,8 @@
+import {Entity} from './Entities'
+
+export class PlayerLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, "torpedo");
+    this.body.velocity.y = -200;
+  }
+}
