+++ src/scenes/SceneGameOver.js
@@ -5,11 +5,11 @@ export class SceneGameOver extends Phaser.Scene {
     super({ key: "SceneGameOver" });
   }
   preload() {
-    this.load.image("sprBtnRestart", "assets/sprBtnRestart.png");
-    this.load.image("sprBtnRestartHover", "assets/sprBtnRestartHover.png");
-    this.load.image("sprBtnRestartDown", "assets/sprBtnRestartDown.png");
-    this.load.image("sprBg0", "assets/sprBg0.png");
-    this.load.image("sprBg1", "assets/sprBg1.png");
+    this.load.image("sprBtnRestart", "../assets/sprBtnRestart.png");
+    this.load.image("sprBtnRestartHover", "../assets/sprBtnRestartHover.png");
+    this.load.image("sprBtnRestartDown", "../assets/sprBtnRestartDown.png");
+    this.load.image("sprBg0", "../assets/sprBg0.png");
+    this.load.image("sprBg1", "../assets/sprBg1.png");
   }
   create() {
