+++ src/scenes/SceneMainMenu.js
@@ -6,13 +6,13 @@ export class SceneMainMenu extends Phaser.Scene {
   }
 
   preload() {
-    this.load.image("sprBg0", "assets/sprBg0.png");
-    this.load.image("sprBg1", "assets/sprBg1.png");
-    this.load.image("sprBtnPlay", "assets/sprBtnPlay.png");
-    this.load.image("sprBtnPlayHover", "assets/sprBtnPlayHover.png");
-    this.load.image("sprBtnPlayDown", "assets/sprBtnPlayDown.png");
-    this.load.audio("sndBtnOver", "assets/sndBtnOver.wav");
-    this.load.audio("sndBtnDown", "assets/sndBtnDown.wav");
+    this.load.image("sprBg0", "../assets/sprBg0.png");
+    this.load.image("sprBg1", "../assets/sprBg1.png");
+    this.load.image("sprBtnPlay", "../assets/sprBtnPlay.png");
+    this.load.image("sprBtnPlayHover", "../assets/sprBtnPlayHover.png");
+    this.load.image("sprBtnPlayDown", "../assets/sprBtnPlayDown.png");
+    this.load.audio("sndBtnOver", "../assets/sndBtnOver.wav");
+    this.load.audio("sndBtnDown", "../assets/sndBtnDown.wav");
   }
 
   create() {
